query_alert/requirements.txt
requests==2.31.0
minio==7.2.9
pillow==10.4.0
